+++ test/LambdaLogger/debug.spec.ts
@@ -1,0 +1,38 @@
+import { LambdaLoggerFactory } from "../../src/index";
+import * as util from "util";
+import { LogLevel } from "../../src/LogLevel";
+import TestUtility from "../lib/TestUtility";
+
+describe("LambdaLogger.debug", () => {
+  it("should be logged output", async () => {
+    const messages = ["hello", "world"];
+
+    const lambdaLogger = LambdaLoggerFactory.create(
+      LogLevel.INFORMATIONAL,
+      TestUtility.extractSlackTokenFromEnv(),
+      TestUtility.extractSlackChannelFromEnv()
+    );
+    const logOutput = await lambdaLogger.debug(messages);
+
+    const expectedContext = `DEBUG \n ${util.inspect(messages, false, null)}`;
+
+    expect(logOutput.logLevel).toBe("DEBUG");
+    expect(logOutput.context).toBe(expectedContext);
+  });
+
+  it("should log is notified to the slack", async () => {
+    const messages = ["hello", "world"];
+
+    const lambdaLogger = LambdaLoggerFactory.create(
+      LogLevel.INFORMATIONAL,
+      TestUtility.extractSlackTokenFromEnv(),
+      TestUtility.extractSlackChannelFromEnv()
+    );
+    const logOutput = await lambdaLogger.debug(messages, true);
+
+    const expectedContext = `DEBUG \n ${util.inspect(messages, false, null)}`;
+
+    expect(logOutput.logLevel).toBe("DEBUG");
+    expect(logOutput.context).toBe(expectedContext);
+  });
+});
